try to generate sdist from pypi packages
While it works for many packages as is, there are many for which it
doesn't yet work. Problems found most often are: Can't find the source
to build the sdist from the pypi metadata. Need to deal with other SCMs,
like hg. Some build binaries even for the sdist and thus require more
dependencies at that time. Some fail on python 3...

This creates OBS packages that create RPMs that when installed put the
pypi sdist source archives in the right place so pip can pick them up
for building offline.

To confgiure pip to use these run:
[global]
no-index = true
find-links = file:///usr/share/python-wheels/
EOF

This can be used with KIWI in OBS to build OpenStack LOCI images.